+++ src/components/index.ts
@@ -1,0 +1,10 @@
+export { default as BaseButton } from './BaseButton.vue'
+export { default as Header } from './Header.vue'
+export { default as ImagePreview } from './ImagePreview.vue'
+export { default as ImageTile } from './ImageTile.vue'
+export { default as InputCheckbox } from './InputCheckbox.vue'
+export { default as InputText } from './InputText.vue'
+export { default as InputTextArea } from './InputTextArea.vue'
+export { default as Modal } from './Modal.vue'
+export { default as WishlistHeader } from './WishlistHeader.vue'
+export { default as WishlistItem } from './WishlistItem.vue'
